requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appengine-python-standard==1.1.2
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.2
cachelib==0.10.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
Deprecated==1.2.14
distlib==0.3.6
Django==3.2.20
django-bootstrap4==23.1
dnspython==2.3.0
filelock==3.12.2
Flask==2.2.5
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
Flask-Session==0.5.0
Flask-SocketIO==5.3.4
Flask-Testing==0.8.1
frozendict==2.3.8
frozenlist==1.3.3
gevent==22.10.2
gevent-websocket==0.10.1
google-auth==2.21.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mock==5.0.2
multidict==6.0.4
openai==0.27.8
packaging==23.1
Pillow==9.5.0
platformdirs==3.8.0
pluggy==1.2.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pymongo==4.4.0
pytest==7.4.0
python-dotenv==0.21.1
python-engineio==4.5.0
python-jwt==4.0.0
python-socketio==5.8.0
pytz==2023.3
requests==2.31.0
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
simple-websocket==0.10.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.23.1
Werkzeug==2.2.3
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
